Update functions.php
I have one of those content upgrade/ opt-in/ get this post via email/ whatever the hell else it's called/ ....forms that allow the person to simply download the post (with an 'XYZ' bonus). So, I needed to add a shortcode that appears in excerpts only.  That way, when a viewer is on an /archive page, they see all the posts and all the optins, yet when they go to the actual post, it disappears (b/c something else happens on the actual post page).  So, I added line 260 with a tid-bit of documentation.  

Merge it or leave it. Your choice. It is now just an option